requirements.txt
torch
transformers
accelerate
datasets
peft
scikit-learn
pandas
numpy
matplotlib
tqdm
pyyaml
streamlit
